+++ postpic/__init__.py
@@ -28,13 +28,14 @@ from . import helper
 from . import datahandling
 from . import particles
 from . import experimental
+from . import plotting
 from .datahandling import *
 from .helper import *
 from .particles import *
 from . import datareader
 from .datareader import chooseCode, readDump, readSim
-from . import plotting
 from ._version import get_versions
+from .io import *
 
 __all__ = ['helper']
 __all__ += datahandling.__all__
@@ -43,6 +44,7 @@ __all__ += particles.__all__
 __all__ += ['datareader', 'plotting']
 # high level functions
 __all__ += ['chooseCode', 'readDump', 'readSim']
+__all__ += io.__all__
 
 __version__ = get_versions()['version']
 __git_version__ = get_versions()['full-revisionid']