+++ postpic/io.py
@@ -15,12 +15,39 @@
 # along with postpic. If not, see <http://www.gnu.org/licenses/>.
 #
 # Stefan Tietze, 2017
+'''
+The postpic.io module provides free functions for importing and exporting data.
+'''
 
 import numpy as np
-import postpic
-from . import datahandling
-import collections
-import pyvtk
+
+__all__ = ['export_field', 'load_field']
+
+
+def load_field(filename):
+    '''
+    construct a new field object from file. currently, the following file
+    formats are supported:
+    *.npz
+    '''
+    if not filename.endswith('npz'):
+        raise ValueError('File format of filename {0} not recognized.'.format(filename))
+    return _import_field_npy(filename)
+
+
+def export_field(filename, field):
+    '''
+    export Field object as a file. Format depends on the extention
+    of the filename. Currently supported are:
+    *.npz
+    *.csv
+    '''
+    if filename.endswith('npz'):
+        _export_field_npy(filename, field)
+    elif filename.endswith('csv'):
+        _export_field_csv(filename, field)
+    else:
+        raise ValueError('File format of filename {0} not recognized.'.format(filename))
 
 
 def _export_field_csv(field, filename):
@@ -28,16 +55,17 @@ def _export_field_csv(field, filename):
     Export the data of a Field object as a CSV file.
     The extent will be given in the comments of that file.
     '''
+    from . import __version__
     if field.dimensions == 1:
         data = np.asarray(field.matrix)
         extent = field.extent
         header = 'Created with postpic version {0}.\nx = [{1}, {2}]'.format(
-            postpic.__version__, extent[0], extent[1])
+            __version__, extent[0], extent[1])
         np.savetxt(filename, data, header=header)
     elif field.dimensions == 2:
         extent = field.extent
         header = 'Created with postpic version {0}.\nx = [{1}, {2}]\ny = [{3}, {4}]'.format(
-            postpic.__version__, extent[0], extent[1], extent[2], extent[3])
+            __version__, extent[0], extent[1], extent[2], extent[3])
         data = np.asarray(field.matrix)
         np.savetxt(filename, data, header=header)
     else:
@@ -90,6 +118,7 @@ def _import_field_npy(filename):
     '''
     import a field object from a file written by _export_field_npy()
     '''
+    from .datahandling import Field, Axis
     import_file = np.load(filename)
 
     # Axes Objects
@@ -102,21 +131,22 @@ def _import_field_npy(filename):
 
     axes = []
     for nax in range(0, len(length_edges)):
-        axes.append(datahandling.Axis(name=meta_ax_names[nax],
-                                      unit=meta_ax_units[nax],
-                                      grid_node=meta_ax_edges[nax, 0:length_edges[nax]]))
+        axes.append(Axis(name=meta_ax_names[nax],
+                         unit=meta_ax_units[nax],
+                         grid_node=meta_ax_edges[nax, 0:length_edges[nax]]))
 
     # field
     meta_field = import_file['meta_field']
-    import_field = datahandling.Field(matrix=import_file['matrix'],
-                                      name=meta_field[0], unit=meta_field[1],
-                                      axes=axes,
-                                      axes_transform_state=meta_ax_transform_state,
-                                      transformed_axes_origins=meta_ax_transformed_origins)
+    import_field = Field(matrix=import_file['matrix'],
+                         name=meta_field[0], unit=meta_field[1],
+                         axes=axes,
+                         axes_transform_state=meta_ax_transform_state,
+                         transformed_axes_origins=meta_ax_transformed_origins)
     import_field.label = meta_field[2]
     import_field.infostring = meta_field[3]
 
     return import_field
+
 
 def export_scalar_vtk(scalarfields, filename):
     if not isinstance(scalarfields, collections.Iterable):
@@ -136,13 +166,14 @@ def export_scalar_vtk(scalarfields, filename):
     else:
         raise Exception('Only 2D or 3D fields are supported.')
 
+    grid = pyvtk.StructuredPoints(dimensions=lengths, origin=starts, spacing=increments)
 
-    grid = pyvtk.StructuredPoints(dimensions=lengths,origin=starts,spacing=increments)
-
-    scalar_list = [pyvtk.Scalars(scalars=np.asarray(f).T.flatten(),name=f.name) for f in scalarfields]
+    scalar_list = []
+    for f in scalarfields:
+        scalar_list.append(pyvtk.Scalars(scalars=np.asarray(f).T.flatten(), name=f.name))
     pointData = pyvtk.PointData(*scalar_list)
 
-    vtk = pyvtk.VtkData(grid,pointData)
+    vtk = pyvtk.VtkData(grid, pointData)
     vtk.tofile(filename)
 
     return