+++ postpic/plotting/plotter_matplotlib.py
@@ -24,6 +24,7 @@ This Class can be used to plot Field Objects using the matplotlib interface.
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import numpy as np
+import warnings
 
 
 __all__ = ['MatplotlibPlotter']